Move deposit/withdraw status into their own file
